New MODULE_API format.
The Module_GetApi (renamed) function returns a pointer to a base structure, where the version in the base structure determines
the content behind the pointer.  In this version of the gateway, the MODULE_API will point to MODULE_API_1, which will have function pointers to CreateFromJson, Create, Start, Destroy and Receive.