Fix JNIEnv* share between start thread and receive thread #275
This commit moves JNIEnv* variable from JAVA_MODULE_HANDLE_DATA structure to local variable.
The variable is assigned by JVM in AttachCurrentThread in both of start and receive functions,
it causes race condition of the variable, then segmentation fault is occurred in ExceptionOccurred JNI function.
So, the JNIEnv* variable should be local instead of a part of JAVA_MODULE_HANDLE_DATA.