Specify libdir when configuring libuv
The CMakeLists.txt in core assumes that libuv is installed in ${edge-root}/build_libuv/dist/lib
but this is not always the case. Some 64bit linux distributions will install libraries to ${prefix}/lib64 instead of ${prefix}/lib.
Fix this by forcing libdir to ${prefix}/lib instead of relying on distro-specific behaviour.