Use per-file UUIDs for TestEndpoint URI
Channel tests are failing on Windows builds in the gate because x64 and x86 jobs are running simultaneously on the same machine, and causing resource conflicts. All Node.js remote module tests which use the TestEndpoint class are connecting to URI 'abc' (which is resolves to a named pipe in nanomsg on Windows), so one test run is stealing the pipe away from the other.

This fix gives each test run (actually, each test class within a run) its own unique URI (a UUID). This eliminates resource contention, but still allows us to test that domain socket files are being cleaned up correctly on Posix systems.