(#474) Retry when nn_send/recv/close/shutdown is interrupted by a signal
If calls to our underlying message broker (nanomsg) are interrupted due to a posix signal (on *nix platforms), we interpret that as an error condition and fail. But it's not an error condition, and can actually be quite common (in fact it's more or less guaranteed if a custom module makes use of posix signals, e.g., for events).

This change adds handling logic after calls to nn_send(), nn_recv(), nn_close(), and nn_shutdown() to detect the interrupted condition (nn function returns -1, errno = EINTR) and retry the command as recommended by the nanomsg documentation.

Note that I didn't bother detecting EINTR when NN_DONTWAIT is specified, because in that case the function will return errno = EAGAIN, which we're already handling.

Note also there are some other nanomsg functions that can return EINTR, but I did an audit of our code and these four are the only ones we're using.