Add definition for CI build (#667)
This change adds the Azure Pipelines build definition that builds IoT Edge v1 and runs all tests, **including end-to-end**, when changes are pushed to master. The build is **not** triggered by pull requests. The pipeline that references this definition lives in the product team's internal DevOps project to protect against test secrets leaking into public-facing logs.

The build runs on Debian 8, Ubuntu 16.04, and Windows Server 2016. The YAML file has a lot in common with iotedge.yml, so it would be good to use job-level templates to reduce duplication. However I don't anticipate this repo will see a lot of churn so I chose not to invest in templates.

Note: I had to upgrade our use of .NET Core to a newer version of 1.x. The one we were referencing isn't available on build agents.